Move version to internal package
