+++ cmd/replicate/main.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/spf13/cobra"
 
+	"github.com/replicate/cli/internal"
 	"github.com/replicate/cli/internal/cmd"
 	"github.com/replicate/cli/internal/cmd/hardware"
 	"github.com/replicate/cli/internal/cmd/model"
@@ -15,7 +16,7 @@ import (
 // rootCmd represents the base command when called without any subcommands
 var rootCmd = &cobra.Command{
 	Use:     "replicate",
-	Version: cmd.Version(),
+	Version: internal.Version(),
 }
 
 // Execute adds all child commands to the root command and sets flags appropriately.
